+++ src/WorldStorage/FireworksSerializer.h
@@ -66,7 +66,7 @@ public:
 	static void WriteToNBTCompound(const cFireworkItem & a_FireworkItem, cFastNBTWriter & a_Writer, const ENUM_ITEM_ID a_Type);
 
 	/** Reads NBT data from a NBT object and populates a FireworkItem with it */
-	static void ParseFromNBT(cFireworkItem & a_FireworkItem, const cParsedNBT & a_NBT, int a_TagIdx, const ENUM_ITEM_ID a_Type);
+	static void ParseFromNBT(cFireworkItem & a_FireworkItem, const cParsedNBT & a_NBT, size_t a_TagIdx, const ENUM_ITEM_ID a_Type);
 
 	/** Converts the firework's vector of colours into a string of values separated by a semicolon */
 	static AString ColoursToString(const cFireworkItem & a_FireworkItem);
